+++ examples/13-elixir-first-liveview/.formatter.exs
@@ -1,0 +1,4 @@
+[
+  plugins: [Phoenix.LiveView.HTMLFormatter],
+  inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"]
+]