+++ examples/14-abstraction-lab/lib/abstraction_lab/exponential_retry_policy.ex
@@ -1,4 +1,7 @@
 defmodule AbstractionLab.ExponentialRetryPolicy do
+  def new() do
+    %{:__reflaxe_class__ => AbstractionLab.ExponentialRetryPolicy}
+  end
   def should_retry(struct, attempt, _last_error) do
     attempt < apply(Map.get(struct, :__reflaxe_class__) || Map.get(struct, :__struct__), :max_attempts, [struct])
   end