+++ examples/16-portable-chat-domain/test/test_helper.exs
@@ -1,0 +1,5 @@
+ExUnit.start()
+
+for file <- Path.wildcard("test/generated/**/*_test.exs") do
+  Code.require_file(file)
+end